Move AudioPlayer out of radioblocks
